refactor: rename snake_case test files in test/config/ to CamelCase
